move tests, add enum description to quoteLength
